+++ pyoffers/models/__init__.py
@@ -5,3 +5,8 @@ from .core import ModelManager
 from .country import Country, CountryManager
 from .goal import Goal, GoalManager
 from .offer import Offer, OfferManager
+
+
+MODEL_MANAGERS = (
+    AdvertiserManager, ConversionManager, CountryManager, GoalManager, OfferManager
+)